content_service/requirements.txt
    fastapi
    boto3
    python-multipart
    pydantic
    uvicorn
    redis
    python-dotenv
    pyjwt